+++ resources/js/types/index.d.ts
@@ -1,9 +1,15 @@
+export interface Role {
+    id: number;
+    name: string;
+}
+
 export interface User {
     id: number;
     name: string;
     username: string;
     email: string;
     email_verified_at?: string;
+    roles?: Role[];
 }
 
 export interface Laboratory {