+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\LaboratoryController;
 use App\Http\Controllers\EquipmentController;
+use App\Http\Controllers\UserManagementController;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
@@ -27,6 +28,7 @@ Route::middleware('auth')->group(function () {
 
     Route::resource('laboratories', LaboratoryController::class);
     Route::resource('equipment', EquipmentController::class);
+    Route::resource('users', UserManagementController::class);
 });
 
 require __DIR__.'/auth.php';